+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## 10.0.0 - 2021-xx-xx
 
 ### Changed
-- **BREAKING**: No changes to the public API. Must use bedrock-kms-http@8,
-  bedrock-meter-http@2, bedrock-meter-usage-reporter@3, bedrock-profile@12,
+- **BREAKING**: No changes to the public API. Must use bedrock-kms-http@9,
+  bedrock-meter-http@2, bedrock-meter-usage-reporter@4, bedrock-profile@12,
   and bedrock-profile-http@8.
 
 ## 9.0.0 - 2021-08-24